requirements.txt
sqlalchemy
pydantic
fastapi
uvicorn
psycopg2-binary
jinja2
loguru
requests
starlette
Pillow
numpy
imagehash
scipy

opencv-python

pip
wheel
setuptools

Mock.GPIO

Shapely
scikit-learn